requirements.txt
streamlitstreamlit-image-coordinatesopencv-python-headlesspillownumpymatplotlibrembg